+++ test/test_gtfs_stops_clustering.rb
@@ -8,9 +8,9 @@ class GtfsStopClustersTest < Minitest::Test
   include GtfsStopsClustering
 
   def test_gtfs_stop_clusters_success
-    gtfs_urls = ["https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"]
+    gtfs_paths = ["test/fixtures/sample-feed.zip"]
 
-    clusters = build_clusters(gtfs_urls, 0.3, 1, 0.85, "path/to/stop_config.txt")
+    clusters = build_clusters(gtfs_paths, 0.3, 1, 0.85)
 
     assert_instance_of Hash, clusters
     clusters.each_value do |cluster|
@@ -21,81 +21,30 @@ class GtfsStopClustersTest < Minitest::Test
     end
   end
 
-  def test_gtfs_paths_nil
-    gtfs_urls = nil
+  def test_correct_clustering
+    gtfs_paths = ["test/fixtures/sample-feed-2.zip"]
+    stops_config_file = "test/fixtures/stops_config_sample.txt"
 
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, 0.3, 1, 0.85, "path/to/stop_config.txt")
+    clusters = build_clusters(gtfs_paths, 0.3, 1, 0.85, stops_config_file)
+
+    assert_instance_of Hash, clusters
+    clusters.each_value do |cluster|
+      assert_instance_of Array, cluster
+      cluster.each do |stop|
+        assert_instance_of Hash, stop
+      end
     end
-  end
 
-  def test_gtfs_paths_not_array
-    gtfs_urls = "https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"
+    stops_not_clustered = clusters[-1]
+    assert !stops_not_clustered.nil?
+    assert_equal 4, stops_not_clustered.length
 
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, 0.3, 1, 0.85, "path/to/stop_config.txt")
-    end
-  end
+    first_cluster = clusters[0]
+    assert !first_cluster.nil?
+    assert_equal 2, first_cluster.length
 
-  def test_gtfs_paths_empty
-    gtfs_urls = []
-
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, 0.3, 1, 0.85, "path/to/stop_config.txt")
-    end
-  end
-
-  def test_epsilon_not_float
-    gtfs_urls = ["https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"]
-    epsilon = "0.3"
-
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, epsilon, 1, 0.85, "path/to/stop_config.txt")
-    end
-  end
-
-  def test_epsilon_negative
-    gtfs_urls = ["https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"]
-    epsilon = -0.3
-
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, epsilon, 1, 0.85, "path/to/stop_config.txt")
-    end
-  end
-
-  def test_min_points_not_integer
-    gtfs_urls = ["https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"]
-    min_points = "1"
-
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, 0.3, min_points, 0.85, "path/to/stop_config.txt")
-    end
-  end
-
-  def test_min_points_negative
-    gtfs_urls = ["https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"]
-    min_points = -1
-
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, 0.3, min_points, 0.85, "path/to/stop_config.txt")
-    end
-  end
-
-  def test_names_similarity_not_float
-    gtfs_urls = ["https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"]
-    names_similarity = "0.85"
-
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, 0.3, 1, names_similarity, "path/to/stop_config.txt")
-    end
-  end
-
-  def test_names_similarity_negative
-    gtfs_urls = ["https://developers.google.com/static/transit/gtfs/examples/sample-feed.zip"]
-    names_similarity = -0.85
-
-    assert_raises ArgumentError do
-      build_clusters(gtfs_urls, 0.3, 1, names_similarity, "path/to/stop_config.txt")
-    end
+    second_cluster = clusters[1]
+    assert !second_cluster.nil?
+    assert_equal 3, second_cluster.length
   end
 end
